Minor Javadoc fixes, unnecessary cast removals.
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/batik/trunk@594379 13f79535-47bb-0310-9956-ffa450edef68

+++ sources/org/apache/batik/gvt/font/SVGGVTGlyphVector.java
@@ -603,7 +603,7 @@ public final class SVGGVTGlyphVector implements GVTGlyphVector {
                 }
                 if (j != glyphs.length) {
                     Glyph bg = glyphs[j];
-                    float rEdge = (float)(currentX + bg.getHorizAdvX());
+                    float rEdge = currentX + bg.getHorizAdvX();
                     for (int k=i; k<j; k++) {
                         g = glyphs[k];
                         g.setTransform(null);
